--- src/main/java/com/example/reframe/controller/HomeController.java
@@ -1,14 +1,0 @@
-package com.example.reframe.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class HomeController {
-
-//	@GetMapping("/")
-//	public String homeRedirect() {
-//	    return "redirect:/deposit/main";
-//	}
-
-}
