+++ src/main/resources/templates/user/fragments/header.html
@@ -30,7 +30,7 @@
 		<div class="content-wrapper">
 			<ul class="sub-menu">
 				<li class="title">예적금</li>
-				<li><a href="#">전체 상품</a></li>
+				<li><a href="/deposit/list">전체 상품</a></li>
 				<li><a href="#">예적금 상품 1</a></li>
 				<li><a href="#">예적금 상품 2</a></li>
 				<li><a href="#">예적금 상품 3</a></li>
@@ -39,15 +39,15 @@
 			</ul>
 			<ul class="sub-menu">
 				<li class="title">카드</li>
-				<li><a href="#">전체 상품</a></li>
-				<li><a href="#">신용카드</a></li>
-				<li><a href="#">기업카드</a></li>
-				<li><a href="#">체크카드</a></li>
+				<li><a href="/card/list">전체 상품</a></li>
+				<li><a href="/card/list?categoryMajor=P">신용카드</a></li>
+				<li><a href="/card/list?categoryMajor=C">기업카드</a></li>
+				<li><a href="/card/list?categoryMajor=K">체크카드</a></li>
 			</ul>
 			<ul class="sub-menu">
 				<li class="title">고객센터</li>
-				<li><a href="#">1:1 문의</a></li>
-				<li><a href="#">자주 묻는 질문</a></li>
+				<li><a href="/qna/qna">1:1 문의</a></li>
+				<li><a href="/faq/faq">자주 묻는 질문</a></li>
 			</ul>
 	    </div>
 	</div>
